Moved map series into a subfolder.
