+++ js/Series/Map/MapSeries.js
@@ -21,20 +21,21 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../Core/Series/Series.js';
-import ColorMapMixin from '../Mixins/ColorMapSeries.js';
+import BaseSeries from '../../Core/Series/Series.js';
+var 
+// indirect dependency to keep product size low
+_a = BaseSeries.seriesTypes, ColumnSeries = _a.column, ScatterSeries = _a.scatter;
+import ColorMapMixin from '../../Mixins/ColorMapSeries.js';
 var colorMapPointMixin = ColorMapMixin.colorMapPointMixin, colorMapSeriesMixin = ColorMapMixin.colorMapSeriesMixin;
-import ColumnSeries from './Column/ColumnSeries.js';
-import H from '../Core/Globals.js';
+import H from '../../Core/Globals.js';
 var noop = H.noop;
-import LegendSymbolMixin from '../Mixins/LegendSymbol.js';
-import LineSeries from './Line/LineSeries.js';
-import mapModule from '../Maps/Map.js';
+import LegendSymbolMixin from '../../Mixins/LegendSymbol.js';
+import LineSeries from '../Line/LineSeries.js';
+import mapModule from '../../Maps/Map.js';
 var maps = mapModule.maps, splitPath = mapModule.splitPath;
-import Point from '../Core/Series/Point.js';
-import ScatterSeries from './Scatter/ScatterSeries.js';
-import SVGRenderer from '../Core/Renderer/SVG/SVGRenderer.js';
-import U from '../Core/Utilities.js';
+import Point from '../../Core/Series/Point.js';
+import SVGRenderer from '../../Core/Renderer/SVG/SVGRenderer.js';
+import U from '../../Core/Utilities.js';
 var extend = U.extend, fireEvent = U.fireEvent, getNestedProperty = U.getNestedProperty, isArray = U.isArray, isNumber = U.isNumber, merge = U.merge, objectEach = U.objectEach, pick = U.pick, splat = U.splat;
 /* *
  *
@@ -1024,7 +1025,7 @@ var MapPoint = /** @class */ (function (_super) {
         series.chart.redraw();
     };
     return MapPoint;
-}(Point));
+}(ScatterSeries.prototype.pointClass));
 MapSeries.prototype.pointClass = MapPoint;
 extend(MapPoint.prototype, colorMapPointMixin);
 BaseSeries.registerSeriesType('map', MapSeries);
