+++ tools/gulptasks/jsdoc-classes.js
@@ -63,7 +63,7 @@ const SOURCE_GLOBS = [
     'js/Series/AreaRangeSeries.js',
     'js/Series/Line/LineSeries.js',
     'js/Series/Column/ColumnSeries.js',
-    'js/Series/MapSeries.js',
+    'js/Series/Map/MapSeries.js',
     'js/Series/OrganizationSeries.js',
     'js/Series/Pie/PieSeries.js',
     'js/Series/PackedBubbleSeries.js',
