+++ ts/Series/MapBubbleSeries.ts
@@ -10,7 +10,7 @@
 
 'use strict';
 
-import type MapSeries from './MapSeries';
+import type MapSeries from './Map/MapSeries';
 import type { SeriesStatesOptions } from '../Core/Series/SeriesOptions';
 import BaseSeries from '../Core/Series/Series.js';
 import Point from '../Core/Series/Point.js';
