+++ ts/Series/MapLineSeries.ts
@@ -12,7 +12,7 @@
 
 import type ColorType from '../Core/Color/ColorType';
 import type LineSeries from './Line/LineSeries';
-import type MapSeries from './MapSeries';
+import type MapSeries from './Map/MapSeries';
 import type { SeriesStatesOptions } from '../Core/Series/SeriesOptions';
 import type { StatesOptionsKey } from '../Core/Series/StatesOptions';
 import type SVGAttributes from '../Core/Renderer/SVG/SVGAttributes';
@@ -60,8 +60,7 @@ declare module '../Core/Series/SeriesType' {
     }
 }
 
-import '../Core/Options.js';
-import './MapSeries.js';
+import './Map/MapSeries.js';
 
 var seriesTypes = BaseSeries.seriesTypes;
